python_data_service/python_data_service/requirements.txt
# Dependências para o serviço de processamento de dados em Python
pymongo
python-dotenv
